Test per-kind flattening; widen the share counters to 32 bits
Flatten already covers every associative-commutative kind (AND, OR,
XOR, BVAND, BVOR, BVXOR, BVPLUS, BVMULT) and merges a same-kind child
only when it is unshared, so flattening never grows the node count.
Pin that per kind: each kind flattens an unshared child, no kind
flattens a shared one, top-level conjuncts flatten irrespective of
sharing, and floating-point chains are left alone.

The share counters in Flatten and Rewriting were uint8_t, so a node
with 257 references wrapped back to a count of one and was treated as
unshared. Widen them to uint32_t; the unordered_map node allocation
does not change size. Regression test included.

+++ include/stp/Simplifier/Rewriting.h
@@ -42,7 +42,8 @@ class Rewriting
   STPMgr* stpMgr;
   NodeFactory* nf;
 
-  std::unordered_map<uint64_t, uint8_t> shareCount;
+  // uint8_t wrapped: 257 references read back as 1, i.e. unshared.
+  std::unordered_map<uint64_t, uint32_t> shareCount;
   std::unordered_map<uint64_t, ASTNode> fromTo;
 
   int removed;
